Split workspace command surface catalogs
