+++ docs/DECISIONS.md
@@ -203,6 +203,7 @@
 - Workspace command IDs reduce through `WorkspaceCommandPlan` and execute through `WorkspaceCommandPlanExecutor` before touching `WorkspaceModel` side effects. The planner owns prefix parsing, draft-prefill mapping, static action mapping, automation quick recurrence parsing, and canonical command-ID-to-tool-name mapping. The executor owns the parsed-plan switch, while `WorkspaceModel` owns the underlying mutations and tool dispatch helpers. This keeps command palette rows, slash-command templates, automation row actions, MCP lifecycle actions, memory actions, and git shortcuts from drifting as the command set grows.
 - Workspace command actions reduce through `WorkspaceCommandActionPlanner` before touching `WorkspaceModel` side effects. The planner owns selected project/thread preconditions, rename draft construction, sidebar bulk command mapping, and action-to-effect routing. `WorkspaceModel` should execute typed `WorkspaceCommandActionEffect` values and keep actor-bound persistence, top-bar refresh, and store synchronization rather than growing another selected-state switch.
 - Command palette surface behavior lives in `WorkspaceCommandPaletteSurface.swift`. Command records, top-bar overflow projection, automation and Computer Use command factories, category ordering, slash/action query scoping, and ranking/scoring stay beside the command contract instead of inside the aggregate `WorkspaceSurface.swift` payload.
+- Command palette row ownership is split by command family. `WorkspaceCommandSurfaceBuilder` composes rows from `WorkspaceThreadCommandCatalog`, `WorkspaceCommandStaticCatalog`, `WorkspaceGitCommandCatalog`, and `WorkspaceProjectCommandCatalog`; it should not grow private command arrays for new features. New command families should get a focused catalog when they carry their own availability rules, keyword derivation, or project/runtime inputs.
 - Automation record construction and due-run draft creation live in `WorkspaceAutomationEngine.swift`, not inside `WorkspaceModel`. `WorkspaceAutomationFactory` owns concrete `QuillAutomation` records and schedule helpers, `WorkspaceAutomationRunner` owns due filtering, recurrence advancement, and generated thread drafts, and `WorkspaceModel` applies those drafts to UI selection, project refresh, stores, and notification reports. This keeps scheduler behavior testable without pulling project persistence or top-bar state into pure automation tests.
 - Automation state mutation lives in `WorkspaceAutomationStateReducer`, not inline in `WorkspaceModel`. Sorting, append/create, status updates, deletion, replacement, and automation-pane visibility changes are pure reducer outputs; the workspace model owns applying the resulting state to persistence and UI side effects.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
